Fix #279. Refactor global object to Guice module
We add InjectedRoutesGenerator to enable dependency injection
and refactor daos and application controllers, so they use DI instead of
relying on Play.current.
Initial data are created as eager singleton by Guice instead of Global.
Sample demonstrates preferable way of creating classes, that need to
execute something on application start.
