--- samples/computer-database/app/Global.scala
@@ -1,651 +1,0 @@
-import java.text.SimpleDateFormat
-
-import scala.concurrent.Await
-import scala.concurrent.duration.Duration
-
-import dao.CompaniesDAO
-import dao.ComputersDAO
-import models.Company
-import models.Computer
-import play.api.Application
-import play.api.GlobalSettings
-
-object Global extends GlobalSettings {
-  override def onStart(app: Application) {
-    InitialData.insert()
-  }
-}
-
-/** Initial set of data to be imported into the sample application. */
-object InitialData {
-  private def companiesDao = new CompaniesDAO
-  private def computersDao = new ComputersDAO
-
-  private val sdf = new SimpleDateFormat("yyyy-MM-dd")
-
-  def insert(): Unit = {
-    val storedComputers = Await.result(computersDao.count(), Duration.Inf)
-    if (storedComputers == 0) {
-      import play.api.libs.concurrent.Execution.Implicits.defaultContext
-      Await.result(companiesDao.insert(Seq(
-        Company(Option(1L), "Apple Inc."),
-        Company(Option(2L), "Thinking Machines"),
-        Company(Option(3L), "RCA"),
-        Company(Option(4L), "Netronics"),
-        Company(Option(5L), "Tandy Corporation"),
-        Company(Option(6L), "Commodore International"),
-        Company(Option(7L), "MOS Technology"),
-        Company(Option(8L), "Micro Instrumentation and Telemetry Systems"),
-        Company(Option(9L), "IMS Associates,  Inc."),
-        Company(Option(10L), "Digital Equipment Corporation"),
-        Company(Option(11L), "Lincoln Laboratory"),
-        Company(Option(12L), "Moore School of Electrical Engineering"),
-        Company(Option(13L), "IBM"),
-        Company(Option(14L), "Amiga Corporation"),
-        Company(Option(15L), "Canon"),
-        Company(Option(16L), "Nokia"),
-        Company(Option(17L), "Sony"),
-        Company(Option(18L), "OQO"),
-        Company(Option(19L), "NeXT"),
-        Company(Option(20L), "Atari"),
-        Company(Option(22L), "Acorn COMPUTER"),
-        Company(Option(23L), "Timex Sinclair"),
-        Company(Option(24L), "Nintendo"),
-        Company(Option(25L), "Sinclair Research Ltd"),
-        Company(Option(26L), "Xerox"),
-        Company(Option(27L), "Hewlett-Packard"),
-        Company(Option(28L), "Zemmix"),
-        Company(Option(29L), "ACVS"),
-        Company(Option(30L), "Sanyo"),
-        Company(Option(31L), "Cray"),
-        Company(Option(32L), "Evans & Sutherland"),
-        Company(Option(33L), "E.S.R. Inc."),
-        Company(Option(34L), "OMRON"),
-        Company(Option(35L), "BBN Technologies"),
-        Company(Option(36L), "Lenovo Group"),
-        Company(Option(37L), "ASUS"),
-        Company(Option(38L), "Amstrad"),
-        Company(Option(39L), "Sun Microsystems"),
-        Company(Option(40L), "Texas Instruments"),
-        Company(Option(41L), "HTC Corporation"),
-        Company(Option(42L), "Research In Motion"),
-        Company(Option(43L), "Samsung Electronics"))), Duration.Inf)
-
-      Await.result(computersDao.insert(Seq(
-        Computer(Option(1L), "MacBook Pro 15.4 inch", None, None, Option(1L)),
-        Computer(Option(2L), "CM-2a", None, None, Option(2L)),
-        Computer(Option(3L), "CM-200", None, None, Option(2L)),
-        Computer(Option(4L), "CM-5e", None, None, Option(2L)),
-        Computer(Option(5L), "CM-5", Option(sdf.parse("1991-01-01")), None, Option(2L)),
-        Computer(Option(6L), "MacBook Pro", Option(sdf.parse("2006-01-10")), None, Option(1L)),
-        Computer(Option(7L), "Apple IIe", None, None, Option(1L)),
-        Computer(Option(8L), "Apple IIc", None, None, Option(1L)),
-        Computer(Option(9L), "Apple IIGS", None, None, Option(1L)),
-        Computer(Option(10L), "Apple IIc Plus", None, None, Option(1L)),
-        Computer(Option(11L), "Apple II Plus", None, None, Option(1L)),
-        Computer(Option(12L), "Apple III", Option(sdf.parse("1980-05-01")), Option(sdf.parse("1984-04-01")), Option(1L)),
-        Computer(Option(13L), "Apple Lisa", None, None, Option(1L)),
-        Computer(Option(14L), "CM-2", None, None, Option(2L)),
-        Computer(Option(15L), "Connection Machine", Option(sdf.parse("1987-01-01")), None, Option(2L)),
-        Computer(Option(16L), "Apple II", Option(sdf.parse("1977-04-01")), Option(sdf.parse("1993-10-01")), Option(1L)),
-        Computer(Option(17L), "Apple III Plus", Option(sdf.parse("1983-12-01")), Option(sdf.parse("1984-04-01")), Option(1L)),
-        Computer(Option(18L), "COSMAC ELF", None, None, Option(3L)),
-        Computer(Option(19L), "COSMAC VIP", Option(sdf.parse("1977-01-01")), None, Option(3L)),
-        Computer(Option(20L), "ELF II", Option(sdf.parse("1977-01-01")), None, Option(4L)),
-        Computer(Option(21L), "Macintosh", Option(sdf.parse("1984-01-24")), None, Option(1L)),
-        Computer(Option(22L), "Macintosh II", None, None, None),
-        Computer(Option(23L), "Macintosh Plus", Option(sdf.parse("1986-01-16")), Option(sdf.parse("1990-10-15")), Option(1L)),
-        Computer(Option(24L), "Macintosh IIfx", None, None, None),
-        Computer(Option(25L), "iMac", Option(sdf.parse("1998-01-01")), None, Option(1L)),
-        Computer(Option(26L), "Mac Mini", Option(sdf.parse("2005-01-22")), None, Option(1L)),
-        Computer(Option(27L), "Mac Pro", Option(sdf.parse("2006-08-07")), None, Option(1L)),
-        Computer(Option(28L), "Power Macintosh", Option(sdf.parse("1994-03-01")), Option(sdf.parse("2006-08-01")), Option(1L)),
-        Computer(Option(29L), "PowerBook", Option(sdf.parse("1991-01-01")), Option(sdf.parse("2006-01-01")), Option(1L)),
-        Computer(Option(30L), "Xserve", None, None, None),
-        Computer(Option(31L), "Powerbook 100", None, None, None),
-        Computer(Option(32L), "Powerbook 140", None, None, None),
-        Computer(Option(33L), "Powerbook 170", None, None, None),
-        Computer(Option(34L), "PowerBook Duo", None, None, None),
-        Computer(Option(35L), "PowerBook 190", None, None, None),
-        Computer(Option(36L), "Macintosh Quadra", Option(sdf.parse("1991-01-01")), None, Option(1L)),
-        Computer(Option(37L), "Macintosh Quadra 900", None, None, None),
-        Computer(Option(38L), "Macintosh Quadra 700", None, None, None),
-        Computer(Option(39L), "Macintosh LC", Option(sdf.parse("1990-01-01")), None, Option(1L)),
-        Computer(Option(40L), "Macintosh LC II", Option(sdf.parse("1990-01-01")), None, Option(1L)),
-        Computer(Option(41L), "Macintosh LC III", Option(sdf.parse("1993-01-01")), None, Option(1L)),
-        Computer(Option(42L), "Macintosh LC III+", None, None, None),
-        Computer(Option(43L), "Macintosh Quadra 605", Option(sdf.parse("1993-10-21")), None, Option(1L)),
-        Computer(Option(44L), "Macintosh LC 500 series", None, None, None),
-        Computer(Option(45L), "TRS-80 Color COMPUTER", Option(sdf.parse("1980-01-01")), None, Option(5L)),
-        Computer(Option(46L), "Acorn System 2", None, None, None),
-        Computer(Option(47L), "Dragon 32/64", None, None, None),
-        Computer(Option(48L), "MEK6800D2", None, None, None),
-        Computer(Option(49L), "Newbear 77/68", None, None, None),
-        Computer(Option(50L), "Commodore PET", None, None, Option(6L)),
-        Computer(Option(51L), "Commodore 64", Option(sdf.parse("1982-08-01")), Option(sdf.parse("1994-01-01")), Option(6L)),
-        Computer(Option(52L), "Commodore 64C", None, None, None),
-        Computer(Option(53L), "Commodore SX-64", None, None, Option(6L)),
-        Computer(Option(54L), "Commodore 128", None, None, Option(6L)),
-        Computer(Option(55L), "Apple I", Option(sdf.parse("1976-04-01")), Option(sdf.parse("1977-10-01")), Option(1L)),
-        Computer(Option(56L), "KIM-1", Option(sdf.parse("1975-01-01")), None, Option(7L)),
-        Computer(Option(57L), "Altair 8800", Option(sdf.parse("1974-12-19")), None, Option(8L)),
-        Computer(Option(58L), "IMSAI 8080", Option(sdf.parse("1975-08-01")), None, Option(9L)),
-        Computer(Option(59L), "IMSAI Series Two", None, None, None),
-        Computer(Option(60L), "VAX", Option(sdf.parse("1977-10-25")), None, Option(10L)),
-        Computer(Option(61L), "VAX 11/780", Option(sdf.parse("1977-10-25")), None, Option(10L)),
-        Computer(Option(62L), "VAX 11/750", Option(sdf.parse("1980-10-01")), None, Option(10L)),
-        Computer(Option(63L), "TX-2", Option(sdf.parse("1958-01-01")), None, Option(11L)),
-        Computer(Option(64L), "TX-0", Option(sdf.parse("1956-01-01")), None, Option(11L)),
-        Computer(Option(65L), "Whirlwind", Option(sdf.parse("1951-04-20")), None, Option(11L)),
-        Computer(Option(66L), "ENIAC", Option(sdf.parse("1946-02-15")), Option(sdf.parse("1955-10-02")), Option(12L)),
-        Computer(Option(67L), "IBM PC", Option(sdf.parse("1981-08-12")), None, Option(13L)),
-        Computer(Option(68L), "Macintosh Classic", None, None, None),
-        Computer(Option(69L), "Macintosh Classic II", Option(sdf.parse("1991-01-01")), None, Option(1L)),
-        Computer(Option(70L), "Amiga", Option(sdf.parse("1985-01-01")), None, Option(14L)),
-        Computer(Option(71L), "Amiga 1000", None, None, Option(6L)),
-        Computer(Option(72L), "Amiga 500", Option(sdf.parse("1987-01-01")), None, Option(6L)),
-        Computer(Option(73L), "Amiga 500+", None, None, None),
-        Computer(Option(74L), "Amiga 2000", Option(sdf.parse("1986-01-01")), Option(sdf.parse("1990-01-01")), Option(6L)),
-        Computer(Option(75L), "Amiga 3000", None, None, Option(6L)),
-        Computer(Option(76L), "Amiga 600", Option(sdf.parse("1992-03-01")), None, Option(6L)),
-        Computer(Option(77L), "Macintosh 128K", Option(sdf.parse("1984-01-01")), None, Option(1L)),
-        Computer(Option(78L), "Macintosh 512K", Option(sdf.parse("1984-09-10")), Option(sdf.parse("1986-04-14")), Option(1L)),
-        Computer(Option(79L), "Macintosh SE", Option(sdf.parse("1987-03-02")), Option(sdf.parse("1989-08-01")), Option(1L)),
-        Computer(Option(80L), "Macintosh SE/30", Option(sdf.parse("1989-01-19")), Option(sdf.parse("1991-10-21")), Option(1L)),
-        Computer(Option(81L), "Canon Cat", Option(sdf.parse("1987-01-01")), None, Option(15L)),
-        Computer(Option(82L), "Nokia 770", None, None, Option(16L)),
-        Computer(Option(83L), "Nokia N800", Option(sdf.parse("2007-01-01")), None, Option(16L)),
-        Computer(Option(84L), "Mylo", Option(sdf.parse("2006-09-21")), None, Option(17L)),
-        Computer(Option(85L), "OQO 02", Option(sdf.parse("2007-01-01")), None, Option(18L)),
-        Computer(Option(86L), "OQO 01+", None, None, None),
-        Computer(Option(87L), "Pinwheel calculator", None, None, None),
-        Computer(Option(88L), "iBook", None, None, Option(1L)),
-        Computer(Option(89L), "MacBook", Option(sdf.parse("2006-05-16")), None, Option(1L)),
-        Computer(Option(90L), "NeXTstation", Option(sdf.parse("1990-01-01")), Option(sdf.parse("1993-01-01")), Option(19L)),
-        Computer(Option(91L), "NeXTcube", Option(sdf.parse("1988-01-01")), Option(sdf.parse("1993-01-01")), Option(19L)),
-        Computer(Option(92L), "NeXTstation Color Turbo", None, None, None),
-        Computer(Option(93L), "NeXTstation Color", None, None, None),
-        Computer(Option(94L), "NeXTstation Turbo", None, None, None),
-        Computer(Option(95L), "NeXTcube Turbo", None, None, Option(19L)),
-        Computer(Option(96L), "NeXTcube 040", None, None, Option(19L)),
-        Computer(Option(97L), "NeXTcube 030", None, None, Option(19L)),
-        Computer(Option(98L), "Tinkertoy Tic-Tac-Toe COMPUTER", None, None, None),
-        Computer(Option(99L), "Z3", None, None, None),
-        Computer(Option(100L), "Z4", None, None, None),
-        Computer(Option(101L), "Z1", None, None, None),
-        Computer(Option(102L), "Z2", None, None, None),
-        Computer(Option(103L), "Wang 2200", Option(sdf.parse("1973-05-01")), None, None),
-        Computer(Option(104L), "Wang VS", None, None, None),
-        Computer(Option(105L), "Wang OIS", None, None, None),
-        Computer(Option(106L), "BBC Micro", None, None, Option(22L)),
-        Computer(Option(107L), "IBM 650", Option(sdf.parse("1953-01-01")), Option(sdf.parse("1962-01-01")), Option(13L)),
-        Computer(Option(108L), "Cray-1", None, None, None),
-        Computer(Option(109L), "Cray-3", None, None, None),
-        Computer(Option(110L), "Cray-2", None, None, None),
-        Computer(Option(111L), "Cray-4", None, None, None),
-        Computer(Option(112L), "Cray X1", None, None, None),
-        Computer(Option(113L), "Cray XD1", None, None, None),
-        Computer(Option(114L), "Cray T3D", Option(sdf.parse("1993-01-01")), None, None),
-        Computer(Option(115L), "Cray T3E", Option(sdf.parse("1995-01-01")), None, None),
-        Computer(Option(116L), "Cray C90", None, None, None),
-        Computer(Option(117L), "Cray T90", None, None, None),
-        Computer(Option(118L), "Cray SV1", None, None, None),
-        Computer(Option(119L), "Cray J90", None, None, None),
-        Computer(Option(120L), "Cray XT3", None, None, None),
-        Computer(Option(121L), "Cray CS6400", None, None, None),
-        Computer(Option(122L), "Atari ST", Option(sdf.parse("1985-01-01")), Option(sdf.parse("1993-01-01")), Option(20L)),
-        Computer(Option(123L), "Amiga 2500", None, None, None),
-        Computer(Option(124L), "Amiga 2500", None, None, Option(6L)),
-        Computer(Option(125L), "Amiga 4000", None, None, Option(6L)),
-        Computer(Option(126L), "Amiga 3000UX", None, None, Option(6L)),
-        Computer(Option(127L), "Amiga 3000T", None, None, Option(6L)),
-        Computer(Option(128L), "Amiga 4000T", None, None, Option(6L)),
-        Computer(Option(129L), "Amiga 1200", Option(sdf.parse("1992-10-01")), Option(sdf.parse("1996-01-01")), Option(6L)),
-        Computer(Option(130L), "Atari 1040 STf", Option(sdf.parse("1986-01-01")), None, None),
-        Computer(Option(131L), "Atari 520 ST", Option(sdf.parse("1985-01-01")), None, None),
-        Computer(Option(132L), "Atari 520 STfm", Option(sdf.parse("1986-01-01")), None, None),
-        Computer(Option(133L), "Atari 1040 STe", Option(sdf.parse("1989-01-01")), None, None),
-        Computer(Option(134L), "Atari MEGA STe", Option(sdf.parse("1991-01-01")), None, None),
-        Computer(Option(135L), "Atari 520 ST+", Option(sdf.parse("1985-01-01")), None, None),
-        Computer(Option(136L), "Atari 520 STm", Option(sdf.parse("1985-01-01")), None, None),
-        Computer(Option(137L), "Atari 130 ST", Option(sdf.parse("1985-01-01")), None, None),
-        Computer(Option(138L), "Atari 260 ST", Option(sdf.parse("1985-01-01")), None, None),
-        Computer(Option(139L), "Atari MEGA ST", Option(sdf.parse("1987-01-01")), None, None),
-        Computer(Option(140L), "Atari 520 STf", Option(sdf.parse("1986-01-01")), None, None),
-        Computer(Option(141L), "Atari 1040 STfm", Option(sdf.parse("1986-01-01")), None, None),
-        Computer(Option(142L), "Atari 2080 ST", Option(sdf.parse("1986-01-01")), None, None),
-        Computer(Option(143L), "Atari 260 ST+", Option(sdf.parse("1985-01-01")), None, None),
-        Computer(Option(144L), "Atari 4160 STe", Option(sdf.parse("1988-01-01")), None, None),
-        Computer(Option(145L), "TRS-80 Color COMPUTER 2", None, None, None),
-        Computer(Option(146L), "TRS-80 Color COMPUTER 3", None, None, None),
-        Computer(Option(147L), "TRS-80 Model 1", Option(sdf.parse("1977-01-01")), None, Option(5L)),
-        Computer(Option(148L), "Timex Sinclair 2068", Option(sdf.parse("1983-11-01")), Option(sdf.parse("1984-04-01")), Option(23L)),
-        Computer(Option(149L), "ZX Spectrum", Option(sdf.parse("1982-01-01")), None, Option(25L)),
-        Computer(Option(150L), "Xerox Star", Option(sdf.parse("1981-01-01")), None, Option(26L)),
-        Computer(Option(151L), "Xerox Alto", None, None, None),
-        Computer(Option(152L), "Acorn Archimedes", None, None, Option(22L)),
-        Computer(Option(153L), "Nintendo Entertainment System", None, None, Option(24L)),
-        Computer(Option(154L), "Super Nintendo Entertainment System", Option(sdf.parse("1991-08-01")), Option(sdf.parse("1999-01-01")), Option(24L)),
-        Computer(Option(155L), "Super Famicom", None, None, None),
-        Computer(Option(156L), "Nintendo GameCube", None, None, Option(24L)),
-        Computer(Option(157L), "Game Boy line", None, None, None),
-        Computer(Option(158L), "PlayStation", Option(sdf.parse("1994-12-03")), None, Option(17L)),
-        Computer(Option(159L), "PlayStation 2", Option(sdf.parse("2000-03-24")), None, Option(17L)),
-        Computer(Option(160L), "Game & Watch", None, None, Option(24L)),
-        Computer(Option(161L), "EDSAC", None, None, None),
-        Computer(Option(162L), "IBM System/4 Pi", None, None, None),
-        Computer(Option(163L), "IBM AP-101", None, None, None),
-        Computer(Option(164L), "IBM TC-1", None, None, None),
-        Computer(Option(165L), "IBM AP-101B", None, None, None),
-        Computer(Option(166L), "IBM AP-101S", None, None, Option(13L)),
-        Computer(Option(167L), "ProLiant", None, None, Option(27L)),
-        Computer(Option(168L), "Http://nepomuk.semanticdesktop.org/xwiki/", None, None, None),
-        Computer(Option(169L), "Sinclair QL", Option(sdf.parse("1984-01-01")), Option(sdf.parse("1986-01-01")), Option(25L)),
-        Computer(Option(170L), "Sinclair ZX81", Option(sdf.parse("1981-01-01")), None, Option(25L)),
-        Computer(Option(171L), "Sinclair ZX80", None, None, Option(25L)),
-        Computer(Option(172L), "Atari 65XE", None, None, Option(20L)),
-        Computer(Option(173L), "Deep Blue", None, None, None),
-        Computer(Option(174L), "Macintosh Quadra 650", None, None, None),
-        Computer(Option(175L), "Macintosh Quadra 610", None, None, None),
-        Computer(Option(176L), "Macintosh Quadra 800", None, None, None),
-        Computer(Option(177L), "Macintosh Quadra 950", None, None, None),
-        Computer(Option(178L), "PowerBook 160", None, None, None),
-        Computer(Option(179L), "PowerBook 145B", None, None, None),
-        Computer(Option(180L), "PowerBook 170", None, None, None),
-        Computer(Option(181L), "PowerBook 145", None, None, None),
-        Computer(Option(182L), "PowerBook G3", None, None, None),
-        Computer(Option(183L), "PowerBook 140", None, None, None),
-        Computer(Option(184L), "Macintosh IIcx", None, None, None),
-        Computer(Option(185L), "Powerbook 180", None, None, None),
-        Computer(Option(186L), "PowerBook G4", None, None, None),
-        Computer(Option(187L), "Macintosh XL", None, None, None),
-        Computer(Option(188L), "PowerBook 100", None, None, None),
-        Computer(Option(189L), "PowerBook 2400c", None, None, None),
-        Computer(Option(190L), "PowerBook 1400", None, None, None),
-        Computer(Option(191L), "Macintosh Quadra 630", None, None, None),
-        Computer(Option(192L), "Macintosh Quadra 660AV", None, None, None),
-        Computer(Option(193L), "Macintosh Quadra 840AV", None, None, None),
-        Computer(Option(194L), "PowerBook 5300", None, None, None),
-        Computer(Option(195L), "PowerBook 3400c", None, None, None),
-        Computer(Option(196L), "Macintosh Color Classic", None, None, None),
-        Computer(Option(197L), "Macintosh 512Ke", None, None, None),
-        Computer(Option(198L), "Macintosh IIsi", None, None, None),
-        Computer(Option(199L), "Macintosh IIx", None, None, None),
-        Computer(Option(200L), "PowerBook 500 series", None, None, None),
-        Computer(Option(201L), "Power Macintosh G3", None, None, None),
-        Computer(Option(202L), "Macintosh IIci", None, None, None),
-        Computer(Option(203L), "iMac G5", Option(sdf.parse("2004-08-31")), None, Option(1L)),
-        Computer(Option(204L), "Power Mac G4", None, None, None),
-        Computer(Option(205L), "Power Macintosh 7100", None, None, None),
-        Computer(Option(206L), "Power Macintosh 9600", None, None, None),
-        Computer(Option(207L), "Power Macintosh 7200", None, None, None),
-        Computer(Option(208L), "Power Macintosh 7300", None, None, None),
-        Computer(Option(209L), "Power Macintosh 8600", None, None, None),
-        Computer(Option(210L), "Power Macintosh 6200", None, None, None),
-        Computer(Option(211L), "Power Macintosh 8100", None, None, None),
-        Computer(Option(212L), "Compact Macintosh", None, None, None),
-        Computer(Option(213L), "Power Macintosh 4400", None, None, None),
-        Computer(Option(214L), "Power Macintosh 9500", None, None, None),
-        Computer(Option(215L), "Macintosh Portable", None, None, None),
-        Computer(Option(216L), "EMac", None, None, None),
-        Computer(Option(217L), "Power Macintosh 7600", None, None, None),
-        Computer(Option(218L), "Power Mac G5", None, None, None),
-        Computer(Option(219L), "Power Macintosh 7500", None, None, None),
-        Computer(Option(220L), "Power Macintosh 6100", None, None, None),
-        Computer(Option(221L), "Power Macintosh 8500", None, None, None),
-        Computer(Option(222L), "Macintosh IIvi", None, None, None),
-        Computer(Option(223L), "Macintosh IIvx", None, None, None),
-        Computer(Option(224L), "IMac G3", None, None, None),
-        Computer(Option(225L), "IMac G4", None, None, None),
-        Computer(Option(226L), "Power Mac G4 Cube", None, None, Option(1L)),
-        Computer(Option(227L), "Intel iMac", None, None, None),
-        Computer(Option(228L), "Deep Thought", None, None, Option(13L)),
-        Computer(Option(229L), "Wii", Option(sdf.parse("2006-11-19")), None, Option(24L)),
-        Computer(Option(230L), "IBM System x", None, None, None),
-        Computer(Option(231L), "IBM System i", Option(sdf.parse("2006-01-01")), None, Option(13L)),
-        Computer(Option(232L), "IBM System z", Option(sdf.parse("2006-01-01")), None, Option(13L)),
-        Computer(Option(233L), "IBM System p", Option(sdf.parse("2000-01-01")), None, Option(13L)),
-        Computer(Option(234L), "LC 575", None, None, None),
-        Computer(Option(235L), "Macintosh TV", None, None, None),
-        Computer(Option(236L), "Macintosh Performa", None, None, None),
-        Computer(Option(237L), "Macintosh II series", None, None, None),
-        Computer(Option(238L), "Power Macintosh 6400", None, None, None),
-        Computer(Option(239L), "Power Macintosh 6500", None, None, None),
-        Computer(Option(240L), "Apple PenLite", None, None, None),
-        Computer(Option(241L), "Wallstreet", None, None, None),
-        Computer(Option(242L), "Twentieth Anniversary Macintosh", None, None, None),
-        Computer(Option(243L), "Power Macintosh 5500", None, None, None),
-        Computer(Option(244L), "iBook G3", None, None, Option(1L)),
-        Computer(Option(245L), "Power Macintosh 5200 LC", None, None, None),
-        Computer(Option(246L), "Power Macintosh 5400", None, None, None),
-        Computer(Option(247L), "CM-1", None, None, None),
-        Computer(Option(248L), "MSX", Option(sdf.parse("1983-01-01")), Option(sdf.parse("1995-01-01")), Option(28L)),
-        Computer(Option(249L), "PlayStation 3", None, None, Option(17L)),
-        Computer(Option(250L), "MSX2", Option(sdf.parse("1986-01-01")), None, Option(29L)),
-        Computer(Option(251L), "MSX2+", Option(sdf.parse("1988-01-01")), None, Option(30L)),
-        Computer(Option(252L), "MSX turbo R", Option(sdf.parse("1990-01-01")), None, None),
-        Computer(Option(253L), "Panasonic FS A1GT", None, None, None),
-        Computer(Option(254L), "Panasonic FS A1ST", None, None, None),
-        Computer(Option(255L), "PDP-11", None, None, Option(10L)),
-        Computer(Option(256L), "PDP-1", None, None, Option(10L)),
-        Computer(Option(257L), "PDP-10", None, None, Option(10L)),
-        Computer(Option(258L), "PDP-8", None, None, Option(10L)),
-        Computer(Option(259L), "PDP-6", None, None, Option(10L)),
-        Computer(Option(260L), "DECSYSTEM-20", None, None, Option(10L)),
-        Computer(Option(261L), "PDP-7", None, None, Option(10L)),
-        Computer(Option(262L), "PDP-5", None, None, Option(10L)),
-        Computer(Option(263L), "PDP-12", None, None, Option(10L)),
-        Computer(Option(264L), "LINC", None, None, Option(10L)),
-        Computer(Option(265L), "PDP-14", None, None, Option(10L)),
-        Computer(Option(266L), "PDP-15", None, None, Option(10L)),
-        Computer(Option(267L), "PDP-16", None, None, Option(10L)),
-        Computer(Option(268L), "Cray X2", Option(sdf.parse("2007-01-01")), None, Option(31L)),
-        Computer(Option(269L), "Cray X-MP", Option(sdf.parse("1982-01-01")), None, Option(31L)),
-        Computer(Option(270L), "Evans & Sutherland ES-1", None, None, Option(32L)),
-        Computer(Option(271L), "Commodore VIC-20", Option(sdf.parse("1980-01-01")), None, Option(6L)),
-        Computer(Option(272L), "PowerBook 150", None, None, None),
-        Computer(Option(273L), "MacBook Air", Option(sdf.parse("2008-01-15")), None, Option(1L)),
-        Computer(Option(274L), "Digi-Comp I", Option(sdf.parse("1963-01-01")), None, Option(33L)),
-        Computer(Option(275L), "Digi-Comp", None, None, None),
-        Computer(Option(276L), "Digi-Comp II", None, None, Option(33L)),
-        Computer(Option(277L), "Manchester Mark I", Option(sdf.parse("1949-01-01")), None, None),
-        Computer(Option(278L), "Small-Scale Experimental Machine", Option(sdf.parse("1948-01-01")), None, None),
-        Computer(Option(279L), "Nintendo 64", None, None, Option(24L)),
-        Computer(Option(280L), "Game Boy Advance", None, None, Option(24L)),
-        Computer(Option(281L), "Game Boy", None, None, Option(24L)),
-        Computer(Option(282L), "Nintendo DS Lite", None, None, Option(24L)),
-        Computer(Option(283L), "Nintendo DS", Option(sdf.parse("2004-01-01")), None, Option(24L)),
-        Computer(Option(284L), "Game Boy Color", None, None, Option(24L)),
-        Computer(Option(285L), "Game Boy Advance SP", None, None, Option(24L)),
-        Computer(Option(286L), "Virtual Boy", None, None, Option(24L)),
-        Computer(Option(287L), "Game Boy Micro", None, None, Option(24L)),
-        Computer(Option(288L), "Roadrunner", None, None, Option(13L)),
-        Computer(Option(289L), "HP 9000", None, None, None),
-        Computer(Option(290L), "OMRON Luna-88K2", None, None, None),
-        Computer(Option(291L), "OMRON Luna-88K", None, None, Option(34L)),
-        Computer(Option(292L), "Motorola series 900", None, None, None),
-        Computer(Option(293L), "Motorola M8120", None, None, None),
-        Computer(Option(294L), "Triton Dolphin System 100", None, None, None),
-        Computer(Option(295L), "BBN TC2000", Option(sdf.parse("1989-08-01")), None, Option(35L)),
-        Computer(Option(296L), "WRT54G", None, None, None),
-        Computer(Option(297L), "ThinkPad", Option(sdf.parse("1992-01-01")), None, Option(36L)),
-        Computer(Option(298L), "Apple Newton", Option(sdf.parse("1993-01-01")), Option(sdf.parse("1998-01-01")), Option(1L)),
-        Computer(Option(299L), "Atanasoff-Berry COMPUTER", Option(sdf.parse("1937-01-01")), None, None),
-        Computer(Option(300L), "Atlas COMPUTER", Option(sdf.parse("1962-01-01")), Option(sdf.parse("1974-01-01")), None),
-        Computer(Option(301L), "ASUS Eee PC 901", None, None, Option(37L)),
-        Computer(Option(302L), "ASUS Eee PC 701", None, None, None),
-        Computer(Option(303L), "IBM 7030", Option(sdf.parse("1961-01-01")), None, Option(13L)),
-        Computer(Option(304L), "System/38", Option(sdf.parse("1979-01-01")), None, Option(13L)),
-        Computer(Option(305L), "System/36", Option(sdf.parse("1983-01-01")), Option(sdf.parse("2000-01-01")), Option(13L)),
-        Computer(Option(306L), "IBM 7090", Option(sdf.parse("1959-01-01")), None, Option(13L)),
-        Computer(Option(307L), "IBM RT", None, None, Option(13L)),
-        Computer(Option(308L), "System/360", Option(sdf.parse("1964-01-01")), None, Option(13L)),
-        Computer(Option(309L), "IBM 801", Option(sdf.parse("1980-01-01")), None, Option(13L)),
-        Computer(Option(310L), "IBM 1401", Option(sdf.parse("1959-01-01")), None, Option(13L)),
-        Computer(Option(311L), "ASCI White", Option(sdf.parse("2001-01-01")), Option(sdf.parse("2006-01-01")), Option(13L)),
-        Computer(Option(312L), "Blue Gene", None, None, Option(13L)),
-        Computer(Option(313L), "ASCI Blue Pacific", Option(sdf.parse("1998-01-01")), None, Option(13L)),
-        Computer(Option(314L), "iPhone", Option(sdf.parse("2007-06-01")), None, Option(1L)),
-        Computer(Option(315L), "Nokia N810", Option(sdf.parse("2007-10-17")), None, Option(16L)),
-        Computer(Option(316L), "EDSAC 2", None, None, None),
-        Computer(Option(317L), "Titan", None, None, None),
-        Computer(Option(318L), "Pilot ACE", None, None, None),
-        Computer(Option(319L), "HP Mini 1000", Option(sdf.parse("2008-10-29")), None, Option(27L)),
-        Computer(Option(320L), "HP 2133 Mini-Note PC", Option(sdf.parse("2008-04-15")), None, Option(27L)),
-        Computer(Option(321L), "Kogan Agora Pro", Option(sdf.parse("2008-12-04")), None, None),
-        Computer(Option(322L), "D-Series Machines", None, None, None),
-        Computer(Option(323L), "ZX Spectrum 48K", Option(sdf.parse("1982-01-01")), None, Option(25L)),
-        Computer(Option(324L), "ZX Spectrum 16K", Option(sdf.parse("1982-01-01")), None, Option(25L)),
-        Computer(Option(325L), "ZX Spectrum 128", Option(sdf.parse("1985-09-01")), None, Option(25L)),
-        Computer(Option(326L), "ZX Spectrum +3", None, None, Option(38L)),
-        Computer(Option(327L), "ZX Spectrum +2", Option(sdf.parse("1986-01-01")), None, Option(38L)),
-        Computer(Option(328L), "ZX Spectrum +2A", Option(sdf.parse("1987-01-01")), None, Option(38L)),
-        Computer(Option(329L), "ZX Spectrum +", Option(sdf.parse("1984-06-01")), None, Option(25L)),
-        Computer(Option(330L), "Acer Extensa", None, None, None),
-        Computer(Option(331L), "Acer Extensa 5220", None, None, None),
-        Computer(Option(332L), "Dell Latitude", None, None, None),
-        Computer(Option(333L), "Toshiba Satellite", None, None, None),
-        Computer(Option(334L), "Timex Sinclair 2048", None, None, Option(23L)),
-        Computer(Option(335L), "Sprinter", None, None, None),
-        Computer(Option(336L), "Timex COMPUTER 2048", None, None, None),
-        Computer(Option(337L), "Pentagon", None, None, None),
-        Computer(Option(338L), "Belle", None, None, None),
-        Computer(Option(339L), "Loki", None, None, Option(25L)),
-        Computer(Option(340L), "Hobbit", None, None, None),
-        Computer(Option(341L), "NeXT COMPUTER", None, None, Option(19L)),
-        Computer(Option(342L), "TRS-80", None, None, None),
-        Computer(Option(343L), "TRS-80 Model 2", Option(sdf.parse("1980-01-01")), None, Option(5L)),
-        Computer(Option(344L), "TRS-80 Model 3", None, None, Option(5L)),
-        Computer(Option(345L), "STacy", Option(sdf.parse("1989-01-01")), None, None),
-        Computer(Option(346L), "ST BOOK", Option(sdf.parse("1990-01-01")), None, None),
-        Computer(Option(347L), "Atari 520 STE", Option(sdf.parse("1989-01-01")), None, None),
-        Computer(Option(348L), "Amiga 2000 Model A", None, None, None),
-        Computer(Option(349L), "Amiga 2000 Model B", None, None, None),
-        Computer(Option(350L), "Amiga 2000 Model C", None, None, None),
-        Computer(Option(351L), "IBM 3270", None, None, None),
-        Computer(Option(352L), "CALDIC", None, None, None),
-        Computer(Option(353L), "Modbook", None, None, None),
-        Computer(Option(354L), "Compaq SystemPro", None, None, None),
-        Computer(Option(355L), "ARRA", None, None, None),
-        Computer(Option(356L), "IBM System Cluster 1350", None, None, None),
-        Computer(Option(357L), "Finite element machine", None, None, None),
-        Computer(Option(358L), "ES7000", None, None, None),
-        Computer(Option(359L), "HP MediaSmart Server", None, None, None),
-        Computer(Option(360L), "HP Superdome", None, None, None),
-        Computer(Option(361L), "IBM Power Systems", Option(sdf.parse("2008-01-01")), None, Option(13L)),
-        Computer(Option(362L), "Oslo Analyzer", None, None, None),
-        Computer(Option(363L), "Microsoft Softcard", None, None, None),
-        Computer(Option(364L), "WITCH", None, None, None),
-        Computer(Option(365L), "Analytical engine", None, None, None),
-        Computer(Option(366L), "EDVAC", None, None, None),
-        Computer(Option(367L), "BINAC", None, None, None),
-        Computer(Option(368L), "Earth Simulator", None, None, None),
-        Computer(Option(369L), "BARK", None, None, None),
-        Computer(Option(370L), "Harvard Mark I", Option(sdf.parse("1944-01-01")), None, Option(13L)),
-        Computer(Option(371L), "ILLIAC IV", None, None, None),
-        Computer(Option(372L), "ILLIAC II", None, None, None),
-        Computer(Option(373L), "ILLIAC III", None, None, None),
-        Computer(Option(374L), "Water integrator", None, None, None),
-        Computer(Option(375L), "CSIRAC", None, None, None),
-        Computer(Option(376L), "System X", None, None, None),
-        Computer(Option(377L), "Harvest", None, None, None),
-        Computer(Option(378L), "ChipTest", None, None, None),
-        Computer(Option(379L), "HiTech", None, None, None),
-        Computer(Option(380L), "Bomba", None, None, None),
-        Computer(Option(381L), "ACE", None, None, None),
-        Computer(Option(382L), "ASCI Red", None, None, None),
-        Computer(Option(383L), "ASCI Thors Hammer", None, None, None),
-        Computer(Option(384L), "ASCI Purple", Option(sdf.parse("2005-01-01")), None, Option(13L)),
-        Computer(Option(385L), "ASCI Blue Mountain", None, None, None),
-        Computer(Option(386L), "Columbia", None, None, None),
-        Computer(Option(387L), "HP Integrity", None, None, None),
-        Computer(Option(388L), "APEXC", None, None, None),
-        Computer(Option(389L), "Datasaab D2", None, None, None),
-        Computer(Option(390L), "BRLESC", None, None, None),
-        Computer(Option(391L), "DYSEAC", None, None, None),
-        Computer(Option(392L), "SSEC", Option(sdf.parse("1948-01-01")), None, Option(13L)),
-        Computer(Option(393L), "Hydra", None, None, None),
-        Computer(Option(394L), "FUJIC", None, None, None),
-        Computer(Option(395L), "RAYDAC", None, None, None),
-        Computer(Option(396L), "Harvard Mark III", None, None, None),
-        Computer(Option(397L), "DATAR", None, None, None),
-        Computer(Option(398L), "ReserVec", None, None, None),
-        Computer(Option(399L), "DASK", None, None, None),
-        Computer(Option(400L), "UTEC", None, None, None),
-        Computer(Option(401L), "DRTE COMPUTER", None, None, None),
-        Computer(Option(402L), "PowerEdge", None, None, None),
-        Computer(Option(403L), "Apple Network Server", None, None, None),
-        Computer(Option(404L), "Goodyear MPP", None, None, None),
-        Computer(Option(405L), "Macintosh 128K technical details", None, None, None),
-        Computer(Option(406L), "Power Macintosh G3", None, None, None),
-        Computer(Option(407L), "CER-10", None, None, None),
-        Computer(Option(408L), "CER-20", None, None, None),
-        Computer(Option(409L), "IBM BladeCenter", Option(sdf.parse("2002-01-01")), None, Option(13L)),
-        Computer(Option(410L), "Wisconsin Integrally Synchronized COMPUTER", None, None, None),
-        Computer(Option(411L), "Amstrad CPC", None, None, Option(38L)),
-        Computer(Option(412L), "Amstrad CPC 6128", None, None, Option(38L)),
-        Computer(Option(413L), "Amstrad CPC 664", None, None, Option(38L)),
-        Computer(Option(414L), "Amstrad CPC 464", None, None, Option(38L)),
-        Computer(Option(415L), "Intergraph", None, None, None),
-        Computer(Option(416L), "Enterprise", None, None, None),
-        Computer(Option(417L), "MTX500", None, None, None),
-        Computer(Option(418L), "Acorn Electron", None, None, None),
-        Computer(Option(419L), "Sony Vaio P", Option(sdf.parse("2009-02-01")), None, Option(17L)),
-        Computer(Option(420L), "VAIO", None, None, Option(17L)),
-        Computer(Option(421L), "Sony Vaio P VGN-P588E/Q", None, None, None),
-        Computer(Option(422L), "Sony Vaio P VGN-P530H/G", None, None, None),
-        Computer(Option(423L), "Sony Vaio P VGN-P530H/W", None, None, None),
-        Computer(Option(424L), "Sony Vaio P VGN-P530H/Q", None, None, None),
-        Computer(Option(425L), "Sony Vaio P VGN-P530H/R", None, None, None),
-        Computer(Option(426L), "Sony Vaio P VGN-P588E/R", None, None, None),
-        Computer(Option(427L), "Sony Vaio P VGN-P598E/Q", None, None, None),
-        Computer(Option(428L), "Timex Sinclair 1000", Option(sdf.parse("1982-07-01")), None, Option(23L)),
-        Computer(Option(429L), "Komputer 2086", None, None, None),
-        Computer(Option(430L), "Galaksija", None, None, None),
-        Computer(Option(431L), "Vector-06C", None, None, None),
-        Computer(Option(432L), "Elektronika BK", None, None, None),
-        Computer(Option(433L), "Sun386i", None, None, Option(39L)),
-        Computer(Option(434L), "Xerox Daybreak", Option(sdf.parse("1985-01-01")), Option(sdf.parse("1989-01-01")), None),
-        Computer(Option(435L), "Xerox NoteTaker", None, None, Option(26L)),
-        Computer(Option(436L), "D4a", Option(sdf.parse("1965-01-01")), None, None),
-        Computer(Option(437L), "LGP-30", None, None, None),
-        Computer(Option(438L), "LGP-21", None, None, None),
-        Computer(Option(439L), "ASUS Eee PC 900", Option(sdf.parse("2008-05-01")), None, Option(37L)),
-        Computer(Option(440L), "Atari TT030", None, None, None),
-        Computer(Option(441L), "Bi Am ZX-Spectrum 48/64", None, None, None),
-        Computer(Option(442L), "Bi Am ZX-Spectrum 128", None, None, None),
-        Computer(Option(443L), "PlayStation Portable", None, None, None),
-        Computer(Option(444L), "MSI Wind Netbook", None, None, None),
-        Computer(Option(445L), "Sharp Mebius NJ70A", Option(sdf.parse("2009-04-21")), None, None),
-        Computer(Option(446L), "HTC Snap", None, None, Option(41L)),
-        Computer(Option(447L), "Commodore Educator 64", None, None, Option(6L)),
-        Computer(Option(448L), "Amiga 1500", None, None, Option(6L)),
-        Computer(Option(449L), "Commodore 65", None, None, Option(6L)),
-        Computer(Option(450L), "Commodore 16", None, None, Option(6L)),
-        Computer(Option(451L), "Commodore CBM-II", None, None, Option(6L)),
-        Computer(Option(452L), "Commodore Plus/4", None, None, Option(6L)),
-        Computer(Option(453L), "Commodore LCD", None, None, Option(6L)),
-        Computer(Option(454L), "Commodore MAX Machine", None, None, Option(6L)),
-        Computer(Option(455L), "Aster CT-80", None, None, None),
-        Computer(Option(456L), "Test", Option(sdf.parse("2009-01-01")), Option(sdf.parse("2009-01-01")), None),
-        Computer(Option(457L), "MSI GX723", None, None, None),
-        Computer(Option(458L), "Eee PC 1000HV", Option(sdf.parse("2009-05-22")), None, None),
-        Computer(Option(459L), "VTech Laser 200", Option(sdf.parse("1983-01-01")), None, None),
-        Computer(Option(460L), "CrunchPad", None, None, None),
-        Computer(Option(461L), "Neo Geo", Option(sdf.parse("1990-01-01")), None, None),
-        Computer(Option(462L), "Sega Mega Drive", None, None, None),
-        Computer(Option(463L), "Sega Master System", None, None, None),
-        Computer(Option(464L), "TurboGrafx-16", None, None, None),
-        Computer(Option(465L), "Sun-3", None, None, Option(39L)),
-        Computer(Option(466L), "Pleiades", None, None, None),
-        Computer(Option(467L), "IBM Sequoia", None, None, None),
-        Computer(Option(468L), "Inves Spectrum 48k plus", None, None, None),
-        Computer(Option(469L), "iPhone 3G", None, None, None),
-        Computer(Option(470L), "iPhone 3GS", None, None, None),
-        Computer(Option(471L), "Beagle Board", None, None, Option(40L)),
-        Computer(Option(472L), "HP nPar", None, None, None),
-        Computer(Option(473L), "MacBook Family", None, None, None),
-        Computer(Option(474L), "Reservisor", None, None, None),
-        Computer(Option(475L), "BladeSystem", None, None, None),
-        Computer(Option(476L), "lenovo thinkpad t60p", None, None, None),
-        Computer(Option(477L), "lenovo thinkpad x200", None, None, Option(36L)),
-        Computer(Option(478L), "lenovo thinkpad t60", None, None, None),
-        Computer(Option(479L), "lenovo thinkpad w700", None, None, None),
-        Computer(Option(480L), "lenovo thinkpad t41", None, None, None),
-        Computer(Option(481L), "lenovo thinkpad z61p", None, None, None),
-        Computer(Option(482L), "lenovo thinkpad x61s", None, None, None),
-        Computer(Option(483L), "lenovo thinkpad t43", None, None, None),
-        Computer(Option(484L), "lenovo thinkpad r400", None, None, None),
-        Computer(Option(485L), "lenovo thinkpad x60s", None, None, None),
-        Computer(Option(486L), "lenovo thinkpad x301", None, None, None),
-        Computer(Option(487L), "lenovo thinkpad t42", None, None, None),
-        Computer(Option(488L), "lenovo thinkpad r61", None, None, None),
-        Computer(Option(489L), "lenovo thinkpad w500", None, None, None),
-        Computer(Option(490L), "lenovo thinkpad sl400", None, None, None),
-        Computer(Option(491L), "lenovo thinkpad x40", None, None, None),
-        Computer(Option(492L), "lenovo thinkpad x200 tablet", None, None, Option(36L)),
-        Computer(Option(493L), "lenovo thinkpad t400s", None, None, None),
-        Computer(Option(494L), "Nokia N900", Option(sdf.parse("2009-10-01")), None, Option(16L)),
-        Computer(Option(495L), "Internet Tablet", None, None, None),
-        Computer(Option(496L), "Meiko Computing Surface", Option(sdf.parse("1986-01-01")), Option(sdf.parse("1993-01-01")), None),
-        Computer(Option(497L), "CS-2", None, None, None),
-        Computer(Option(498L), "IBM 701", Option(sdf.parse("1952-01-01")), None, Option(13L)),
-        Computer(Option(499L), "IBM 5100", Option(sdf.parse("1975-01-01")), None, Option(13L)),
-        Computer(Option(500L), "AN/FSQ-7", Option(sdf.parse("1958-01-01")), None, Option(13L)),
-        Computer(Option(501L), "AN/FSQ-32", Option(sdf.parse("1960-01-01")), None, Option(13L)),
-        Computer(Option(502L), "IBM CPC", Option(sdf.parse("1949-01-01")), None, Option(13L)),
-        Computer(Option(503L), "System/34", Option(sdf.parse("1978-01-01")), Option(sdf.parse("1983-01-01")), Option(13L)),
-        Computer(Option(504L), "System/32", Option(sdf.parse("1975-01-01")), None, Option(13L)),
-        Computer(Option(505L), "System/3", Option(sdf.parse("1969-01-01")), Option(sdf.parse("1985-01-01")), Option(13L)),
-        Computer(Option(506L), "IBM 305", Option(sdf.parse("1956-01-01")), None, Option(13L)),
-        Computer(Option(507L), "English Electric DEUCE", None, None, None),
-        Computer(Option(508L), "CER-203", None, None, None),
-        Computer(Option(509L), "CER-22", None, None, None),
-        Computer(Option(510L), "Kentucky Linux Athlon Testbed", None, None, None),
-        Computer(Option(511L), "QNAP TS-101", None, None, None),
-        Computer(Option(512L), "iPad", Option(sdf.parse("2010-01-01")), Option(sdf.parse("2011-03-02")), Option(1L)),
-        Computer(Option(513L), "iPhone 2G", None, None, None),
-        Computer(Option(514L), "Inslaw", None, None, None),
-        Computer(Option(515L), "WePad", Option(sdf.parse("2010-07-01")), None, None),
-        Computer(Option(516L), "MacBook Parts", None, None, Option(1L)),
-        Computer(Option(517L), "MacBook 13-inch Core 2 Duo 2.13GHz (MC240LL/A) DDR2 Model", None, None, Option(1L)),
-        Computer(Option(518L), "MacBook 13-inch Core 2 Duo 2.13GHz (MC240T/A) DDR2 Model", None, None, None),
-        Computer(Option(519L), "MacBook 13-inch Core 2 Duo 2.13GHz (MC240X/A) DDR2 Model", None, None, None),
-        Computer(Option(520L), "MacBook 13-inch Core 2 Duo 2.26GHz (Unibody MC207LL/A) DDR3 Model", None, None, None),
-        Computer(Option(521L), "MC240LL/A", None, None, None),
-        Computer(Option(522L), "D.K.COMMUNICATION", None, None, None),
-        Computer(Option(523L), "iPhone 4", None, None, Option(1L)),
-        Computer(Option(524L), "Nintendo 3DS", Option(sdf.parse("2010-03-23")), None, Option(24L)),
-        Computer(Option(525L), "ASUS Eee PC 1005PE", Option(sdf.parse("2010-01-01")), None, Option(37L)),
-        Computer(Option(526L), "National Law Enforcement System", None, None, None),
-        Computer(Option(527L), "BlackBerry PlayBook", None, None, Option(42L)),
-        Computer(Option(528L), "Barnes & Noble nook", Option(sdf.parse("2009-10-20")), None, None),
-        Computer(Option(529L), "SAM Coupé", None, None, None),
-        Computer(Option(530L), "HTC Dream", Option(sdf.parse("2008-10-22")), None, Option(41L)),
-        Computer(Option(531L), "Samsung Galaxy Tab", Option(sdf.parse("2010-09-02")), None, Option(43L)),
-        Computer(Option(532L), "BlackBerry PlayBook", Option(sdf.parse("2010-09-27")), None, Option(42L)),
-        Computer(Option(533L), "Tianhe-I", None, None, None),
-        Computer(Option(534L), "Kno", None, None, None),
-        Computer(Option(535L), "ThinkPad 701 C", None, None, None),
-        Computer(Option(536L), "ThinkPad 340 CSE", None, None, None),
-        Computer(Option(537L), "ThinkPad 755 CX", None, None, None),
-        Computer(Option(538L), "ThinkPad 755 CE", None, None, None),
-        Computer(Option(539L), "ThinkPad 370 C", None, None, None),
-        Computer(Option(540L), "Coleco Adam", Option(sdf.parse("1983-01-01")), None, None),
-        Computer(Option(541L), "Nebulae", None, None, None),
-        Computer(Option(542L), "Alex eReader", None, None, None),
-        Computer(Option(543L), "Acer Iconia", None, None, None),
-        Computer(Option(544L), "Archos 101", None, None, None),
-        Computer(Option(545L), "Fujitsu Lifebook T900", None, None, None),
-        Computer(Option(546L), "Motorola Xoom", None, None, None),
-        Computer(Option(547L), "ViewSonic G Tablet", None, None, None),
-        Computer(Option(548L), "DEC Professional", Option(sdf.parse("1982-01-01")), None, Option(10L)),
-        Computer(Option(549L), "DEC Multia", Option(sdf.parse("1994-11-07")), None, Option(10L)),
-        Computer(Option(550L), "DEC Firefly", None, None, Option(10L)),
-        Computer(Option(551L), "DEC 3000 AXP", None, None, Option(10L)),
-        Computer(Option(552L), "DEC 2000 AXP", Option(sdf.parse("1993-05-25")), None, Option(10L)),
-        Computer(Option(553L), "DEC 4000 AXP", Option(sdf.parse("1992-11-10")), None, Option(10L)),
-        Computer(Option(554L), "DEC 7000/10000 AXP", Option(sdf.parse("1992-11-10")), None, Option(10L)),
-        Computer(Option(555L), "DEC Professional 350", None, None, None),
-        Computer(Option(556L), "DEC Rainbow 100", None, None, None),
-        Computer(Option(557L), "DEC Professional 325", None, None, None),
-        Computer(Option(558L), "DECmate II", None, None, Option(10L)),
-        Computer(Option(559L), "DECmate", None, None, Option(10L)),
-        Computer(Option(560L), "DECsystem", None, None, Option(10L)),
-        Computer(Option(561L), "NetApp Filer", None, None, None),
-        Computer(Option(562L), "DEC GT40", None, None, Option(10L)),
-        Computer(Option(563L), "ecoATM", None, None, None),
-        Computer(Option(564L), "MindWave BrainCubed Education Bundle", None, None, None),
-        Computer(Option(565L), "PalmPilot", None, None, None),
-        Computer(Option(566L), "Upcoming iPhone 5", None, None, Option(1L)),
-        Computer(Option(567L), "Dell Inspiron 560 Desktop COMPUTER ", None, None, None),
-        Computer(Option(568L), "IPad 2", None, None, Option(1L)),
-        Computer(Option(569L), "HP TouchPad", Option(sdf.parse("2011-02-09")), None, Option(27L)),
-        Computer(Option(570L), "HP Veer", Option(sdf.parse("2011-02-09")), None, Option(27L)),
-        Computer(Option(571L), "Lenovo Thinkpad Edge 11", None, None, Option(36L)),
-        Computer(Option(572L), "Dell Vostro", None, None, None),
-        Computer(Option(573L), "Gateway LT3103U", Option(sdf.parse("2008-01-01")), None, None),
-        Computer(Option(574L), "iPhone 4S", Option(sdf.parse("2011-10-14")), None, Option(1L)))), Duration.Inf)
-    }
-  }
-}
