+++ samples/computer-database/app/bootstrap/ComputersDatabaseModule.scala
@@ -1,0 +1,9 @@
+package bootstrap
+
+import com.google.inject.AbstractModule
+
+class ComputersDatabaseModule extends AbstractModule {
+  protected def configure: Unit = {
+    bind(classOf[InitialData]).asEagerSingleton()
+  }
+}
