+++ samples/computer-database/build.sbt
@@ -1,3 +1,5 @@
 name := "computer-database-sample"
 
 PlayKeys.playOmnidoc := false
+
+routesGenerator := InjectedRoutesGenerator
