+++ samples/computer-database/conf/application.conf
@@ -8,6 +8,8 @@ application.name=computer-database
 # If you deploy your application to several instances be sure to use the same key!
 application.secret="E27D^[_<Lpt0vjad]de;3;tx3gpRmG4ByofnahOIo9gbsMWut1w3xg[>9W"
 
+play.modules.enabled += "bootstrap.ComputersDatabaseModule"
+
 # Database configuration
 # ~~~~~ 
 # You can declare as many datasources as you want.
